requirements.txt
openai>=1.60.0
python-dotenv>=1.0.1
streamlit>=1.42.0
